Move `MapClone` into `Methods` lint pass
